+++ src/concurrency/lease.ts
@@ -2,7 +2,7 @@
  * An acquired lock or resource
  */
 export interface Lease<T> {
-  value: T;
+  readonly value: T;
   dispose(): PromiseLike<void>;
   [Symbol.asyncDispose](): PromiseLike<void>;
 }
